Fixed Service Manifest not visible on GCC
- The "add_compiler_export_flags" command of the GenerateExportHeaders
Module is problematic, because on gcc/llvm it sets the general
visibility of library functions to hidden, to reduce export table size.
This means, it hides the pocoBuildManifest function, which lacks a
visibility declaration in POCO_LIBRARY_API. On windows it has no effect
anyway, so I commented it out in the ServiceLibrary CMake file. Maybe
write a feature request to appinf to make the macro more flexible like
described here: http://gcc.gnu.org/wiki/Visibility
- cleanup of ServiceLibrary CMakeLists to improve readability and
separation between public and private headers.